# openapi.yaml
openapi: 3.0.3
info:
  title: The Vanity E-commerce API
  description: |
    RESTful API for The Vanity E-commerce Platform.
    This API provides endpoints for managing products, users, orders, and admin operations.
  version: 1.0.0
  contact:
    name: The Vanity Support
    email: support@thevanity.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.thevanity.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
  schemas:
    Product:
      type: object
      required:
        - name
        - brand
        - price
        - image
        - category
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Crystal Pendant Necklace Set"
        brand:
          type: string
          example: "Fashion Jewelry"
        price:
          type: number
          example: 600
        originalPrice:
          type: number
          example: 899
        image:
          type: string
          format: uri
          example: "https://example.com/image.jpg"
        images:
          type: array
          items:
            type: string
            format: uri
        category:
          type: string
          example: "Necklaces"
        description:
          type: string
          example: "Elegant crystal pendant necklace set with intricate detailing"
        reviews:
          type: integer
          default: 0
          example: 120
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          default: 0
          example: 4.5
        stock:
          type: integer
          default: 0
          example: 45
        specifications:
          type: object
          additionalProperties: true
        isPublic:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    User:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        firstName:
          type: string
          example: "Priya"
        lastName:
          type: string
          example: "Sharma"
        email:
          type: string
          format: email
          example: "priya.sharma@email.com"
        phone:
          type: string
          example: "+91 98765 43210"
        role:
          type: string
          enum: [user, admin]
          default: user
        status:
          type: string
          enum: [active, suspended]
          default: active
        emailVerified:
          type: boolean
          default: false
        phoneVerified:
          type: boolean
          default: false
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
    Address:
      type: object
      required:
        - firstName
        - lastName
        - address1
        - city
        - state
        - postalCode
        - country
      properties:
        _id:
          type: string
        type:
          type: string
          enum: [shipping, billing]
          default: shipping
        firstName:
          type: string
          example: "Priya"
        lastName:
          type: string
          example: "Sharma"
        address1:
          type: string
          example: "123 MG Road"
        address2:
          type: string
        city:
          type: string
          example: "Mumbai"
        state:
          type: string
          example: "Maharashtra"
        postalCode:
          type: string
          example: "400001"
        country:
          type: string
          example: "India"
        phone:
          type: string
          example: "+91 98765 43210"
        isDefault:
          type: boolean
          default: false
    Order:
      type: object
      required:
        - orderNumber
        - userId
        - items
        - totalAmount
      properties:
        _id:
          type: string
        orderNumber:
          type: string
          example: "ORD-2024-001"
        userId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
        totalAmount:
          type: number
          example: 1200
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
          default: pending
        shippingAddress:
          $ref: '#/components/schemas/Address'
        paymentMethod:
          type: string
          enum: [cod, card, upi, netbanking]
        paymentStatus:
          type: string
          enum: [pending, completed, failed, refunded]
        trackingNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        statusCode:
          type: integer
          example: 400
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve a list of all products with optional filtering
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category name
        - name: limit
          in: query
          schema:
            type: integer
          description: Limit number of results
        - name: exclude
          in: query
          schema:
            type: string
          description: Exclude product by ID
        - name: search
          in: query
          schema:
            type: string
          description: Search in product name, brand, category
        - name: minPrice
          in: query
          schema:
            type: number
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
          description: Maximum price filter
        - name: brand
          in: query
          schema:
            type: string
          description: Filter by brand
        - name: isPublic
          in: query
          schema:
            type: boolean
            default: true
          description: Filter by public/private status
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get single product
      description: Retrieve a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== CATEGORIES ====================
...
  /users/profile:
    get:
      tags:
        - User
      summary: Get user profile
      description: Retrieve authenticated user's profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - User
      summary: Update user profile
      description: Update authenticated user's profile information
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==================== ORDERS ====================
  /orders/my-orders:
    get:
      tags:
        - Orders
      summary: Get user orders
      description: Retrieve all orders for authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders:
    post:
      tags:
        - Orders
      summary: Create order
      description: Create a new order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
                - shippingAddress
                - paymentMethod
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                shippingAddress:
                  $ref: '#/components/schemas/Address'
                paymentMethod:
                  type: string
                  enum: [cod, card, upi, netbanking]
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  # ==================== CART ====================
...
  /admin/products:
    post:
      tags:
        - Admin - Products
      summary: Create product
      description: Create a new product (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /admin/products/{id}:
    put:
      tags:
        - Admin - Products
      summary: Update product
      description: Update existing product (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated successfully

    delete:
      tags:
        - Admin - Products
      summary: Delete product
      description: Delete a product (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully

  # ==================== ADMIN - INVENTORY ====================
...
  /admin/orders:
    get:
      tags:
        - Admin - Orders
      summary: Get all orders (Admin)
      description: Retrieve all orders with optional filtering
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
...